+++ packages/vendure-plugin-accept-blue/src/api/accept-blue-common-resolvers.ts
@@ -13,7 +13,9 @@ import {
 } from '../types';
 import { AcceptBlueService } from './accept-blue-service';
 import {
+  AcceptBluePaymentMethodQuote,
   AcceptBlueSubscription,
+  AcceptBlueSurcharges,
   Query as GraphqlQuery,
   QueryPreviewAcceptBlueSubscriptionsArgs,
   QueryPreviewAcceptBlueSubscriptionsForProductArgs,
@@ -106,5 +108,19 @@ export class AcceptBlueCommonResolver {
     return value.payment_method_type === 'card'
       ? 'AcceptBlueCardPaymentMethod'
       : 'AcceptBlueCheckPaymentMethod';
+  }
+
+  @Query()
+  async eligibleAcceptBluePaymentMethods(
+    @Ctx() ctx: RequestContext
+  ): Promise<AcceptBluePaymentMethodQuote[]> {
+    return this.acceptBlueService.getEligiblePaymentMethods(ctx);
+  }
+
+  @Query()
+  async acceptBlueSurcharges(
+    @Ctx() ctx: RequestContext
+  ): Promise<AcceptBlueSurcharges> {
+    return await this.acceptBlueService.getSurcharges(ctx);
   }
 }