+++ packages/vendure-plugin-accept-blue/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.3.0 (2025-02-06)
+
+- Allow disabling certain payment methods, e.g. 'visa' or 'check', via the Vendure Admin UI
+- Expose surcharges used by Accept Blue via GraphQL query
+
 # 2.2.0 (2025-01-14)
 
 - Update a customer's shipping and billing details in Accept Blue on subscription creation